extract map entities + remove server from model
