+++ src/code/dialogs/keysList.js
@@ -7,7 +7,7 @@ import KeyListPortal from "./keyListPortal";
 import { getSelectedOperation } from "../selectedOp";
 import wX from "../wX";
 
-import PortalUI from "../ui/portal";
+import * as PortalUI from "../ui/portal";
 
 const KeysList = WDialog.extend({
   statics: {
