+++ src/code/dialogs/linkDialog.js
@@ -3,7 +3,7 @@ import WasabeePortal from "../model/portal";
 import { getSelectedOperation } from "../selectedOp";
 import wX from "../wX";
 
-import PortalUI from "../ui/portal";
+import * as PortalUI from "../ui/portal";
 import { displayError } from "../error";
 
 const LinkDialog = WDialog.extend({
