+++ src/code/dialogs/linkListDialog.js
@@ -3,8 +3,8 @@ import wX from "../wX";
 import { loadFaked } from "../uiCommands";
 import { getSelectedOperation } from "../selectedOp";
 
-import PortalUI from "../ui/portal";
-import LinkUI from "../ui/link";
+import * as PortalUI from "../ui/portal";
+import * as LinkUI from "../ui/link";
 
 const LinkListDialog = OperationChecklistDialog.extend({
   statics: {
