+++ src/code/dialogs/sendTargetDialog.js
@@ -1,14 +1,13 @@
 import { WDialog } from "../leafletClasses";
 import WasabeeMarker from "../model/marker";
 import WasabeePortal from "../model/portal";
-import WasabeeMe from "../model/me";
-import WasabeeTeam from "../model/team";
 import { targetPromise } from "../server";
 import wX from "../wX";
 import { getSelectedOperation } from "../selectedOp";
 
-import PortalUI from "../ui/portal";
+import * as PortalUI from "../ui/portal";
 import { displayError, displayInfo } from "../error";
+import { getTeam, getMe } from "../model/cache";
 
 const SendTargetDialog = WDialog.extend({
   statics: {
@@ -95,12 +94,12 @@ const SendTargetDialog = WDialog.extend({
       this._value = menu.value;
     });
 
-    const me = await WasabeeMe.waitGet();
+    const me = await getMe();
     for (const t of operation.teamlist) {
       if (me.teamJoined(t.teamid) == false) continue;
       try {
         // allow teams to be 5 minutes cached
-        const tt = await WasabeeTeam.get(t.teamid, 5 * 60);
+        const tt = await getTeam(t.teamid, 5 * 60);
         for (const a of tt.agents) {
           if (!alreadyAdded.includes(a.id)) {
             alreadyAdded.push(a.id);
