+++ src/code/dialogs/setCommentDialog.js
@@ -4,8 +4,8 @@ import WasabeeLink from "../model/link";
 import WasabeeMarker from "../model/marker";
 import wX from "../wX";
 
-import PortalUI from "../ui/portal";
-import LinkUI from "../ui/link";
+import * as PortalUI from "../ui/portal";
+import * as LinkUI from "../ui/link";
 
 export const SetCommentDialog = WDialog.extend({
   statics: {
