+++ src/code/dialogs/stateDialog.js
@@ -4,8 +4,8 @@ import WasabeeMarker from "../model/marker";
 import wX from "../wX";
 import { getSelectedOperation } from "../selectedOp";
 
-import PortalUI from "../ui/portal";
-import LinkUI from "../ui/link";
+import * as PortalUI from "../ui/portal";
+import * as LinkUI from "../ui/link";
 
 const StateDialog = WDialog.extend({
   statics: {
