+++ src/code/map/agent.d.ts
@@ -1,0 +1,14 @@
+import WasabeeAgent from "../model/agent";
+
+interface WLAgentOptions extends L.MarkerOptions {
+  agent: WasabeeAgent;
+  id: string;
+  zoom: number;
+}
+
+export class WLAgent extends L.Marker {
+  options: WLAgentOptions;
+  constructor(agent: WasabeeAgent);
+  update(): this;
+  _getPopup(): HTMLDivElement;
+}
