+++ src/code/ui/agent.d.ts
@@ -6,16 +6,3 @@ export function formatDisplay(
 ): HTMLAnchorElement;
 
 export function timeSinceformat(agent: WasabeeAgent): string;
-
-interface WLAgentOptions extends L.MarkerOptions {
-  agent: WasabeeAgent;
-  id: string;
-  zoom: number;
-}
-
-export class WLAgent extends L.Marker {
-  options: WLAgentOptions;
-  constructor(agent: WasabeeAgent);
-  update(): this;
-  _getPopup(): HTMLDivElement;
-}
