+++ src/code/ui/portal.js
@@ -1,12 +1,8 @@
 import WasabeePortal from "../model/portal";
 
-import { getSelectedOperation } from "../selectedOp";
-import AssignDialog from "../dialogs/assignDialog";
-import SendTargetDialog from "../dialogs/sendTargetDialog";
-import SetCommentDialog from "../dialogs/setCommentDialog";
 import wX from "../wX";
 
-function fromIITC(p) {
+export function fromIITC(p) {
   // we have all the details
   if (p && p.options && p.options.data && p.options.guid) {
     const data = p.options.data;
@@ -32,19 +28,19 @@ function fromIITC(p) {
   return null;
 }
 
-function team(portal) {
+export function team(portal) {
   if (window.portals[portal.id] && window.portals[portal.id].options.data)
     return window.portals[portal.id].options.data.team;
   return "";
 }
 
-function displayName(portal) {
+export function displayName(portal) {
   if (portal.pureFaked) return wX("FAKED", { portalId: portal.id });
   if (portal.loading) return wX("LOADING1", { portalGuid: portal.id });
   return portal.name;
 }
 
-function displayFormat(portal, shortName = false) {
+export function displayFormat(portal, shortName = false) {
   const pt = portal.latLng;
   const v = `${portal.lat},${portal.lng}`;
   const name = displayName(portal);
@@ -87,147 +83,10 @@ function displayFormat(portal, shortName = false) {
   return e;
 }
 
-function get(id) {
+export function get(id) {
   return fromIITC(window.portals[id]);
 }
 
-function getSelected() {
+export function getSelected() {
   return window.selectedPortal ? get(window.selectedPortal) : null;
 }
-
-// common part for marker and anchors
-const WLPortal = L.Marker.extend({
-  type: "portal",
-
-  initialize: function (options) {
-    const operation = getSelectedOperation();
-    const portal = operation.getPortal(options.portalId);
-    options.title = portal.name;
-    L.Marker.prototype.initialize.call(this, portal.latLng, options);
-    this.bindPopup((layer) => layer._popupContent(), {
-      className: "wasabee-popup",
-      closeButton: false,
-    });
-
-    this.off("click", this._openPopup);
-    window.registerMarkerForOMS(this);
-    this.on("spiderfiedclick", this._openPopup);
-  },
-
-  _popupContent: function () {
-    const div = L.DomUtil.create("div", `wasabee-${this.type}-popup`);
-    return div;
-  },
-
-  _popupPortalComments: function (container, portal, canWrite) {
-    const portalComment = L.DomUtil.create(
-      "div",
-      "wasabee-portal-comment",
-      container
-    );
-    const pcLink = L.DomUtil.create("a", null, portalComment);
-    pcLink.textContent = portal.comment || wX("SET_PORTAL_COMMENT");
-    if (canWrite) {
-      pcLink.href = "#";
-      L.DomEvent.on(pcLink, "click", this._setPortalComment, this);
-    }
-    if (portal.hardness) {
-      const portalHardness = L.DomUtil.create(
-        "div",
-        "wasabee-portal-hardness",
-        container
-      );
-      const phLink = L.DomUtil.create("a", null, portalHardness);
-      phLink.textContent = portal.hardness;
-      if (canWrite) {
-        phLink.href = "#";
-        L.DomEvent.on(phLink, "click", this._setPortalComment, this);
-      }
-    }
-  },
-
-  _setPortalComment: function (ev) {
-    L.DomEvent.stop(ev);
-    const operation = getSelectedOperation();
-    const portal = operation.getPortal(this.options.portalId);
-    const scd = new SetCommentDialog({
-      target: portal,
-      operation: operation,
-    });
-    scd.enable();
-    this.closePopup();
-  },
-
-  _assignButton: function (container, text, target) {
-    const assignButton = L.DomUtil.create("button", null, container);
-    assignButton.textContent = text;
-    L.DomEvent.on(assignButton, "click", (ev) => {
-      L.DomEvent.stop(ev);
-      const ad = new AssignDialog({ target: target });
-      ad.enable();
-      this.closePopup();
-    });
-  },
-
-  _deleteButton: function (container, text) {
-    const deleteButton = L.DomUtil.create("button", null, container);
-    deleteButton.textContent = text;
-    L.DomEvent.on(deleteButton, "click", (ev) => {
-      L.DomEvent.stop(ev);
-      if (this._deleteAction) this._deleteAction();
-      this.closePopup();
-    });
-  },
-
-  _sendTargetButton: function (container, text, target) {
-    const sendButton = L.DomUtil.create("button", null, container);
-    sendButton.textContent = text;
-    L.DomEvent.on(sendButton, "click", (ev) => {
-      L.DomEvent.stop(ev);
-      const std = new SendTargetDialog({ target: target });
-      std.enable();
-      this.closePopup();
-    });
-  },
-
-  _mapButton: function (container, text) {
-    const gmapButton = L.DomUtil.create("button", null, container);
-    gmapButton.textContent = text;
-    L.DomEvent.on(gmapButton, "click", (ev) => {
-      L.DomEvent.stop(ev);
-      this.closePopup();
-      const latLng = this.getLatLng();
-      // use intent on android
-      if (
-        typeof window.android !== "undefined" &&
-        window.android &&
-        window.android.intentPosLink
-      ) {
-        window.android.intentPosLink(
-          +latLng.lat,
-          +latLng.lng,
-          window.map.getZoom(),
-          this.options.title,
-          true
-        );
-      } else {
-        window.open(
-          "https://www.google.com/maps/search/?api=1&query=" +
-            latLng.lat +
-            "," +
-            latLng.lng
-        );
-      }
-    });
-  },
-});
-
-export default {
-  fromIITC,
-  displayName,
-  displayFormat,
-  get,
-  getSelected,
-  team,
-  WLPortal,
-};
